refactor: extract sandbox command wrappers
